added : bot command protection

+++ src/commands/trunc.ts
@@ -21,15 +21,30 @@ export default new Command(
   async (slash) => {
     try {
       const amount = slash.options.getInteger("amount", true);
-      await (slash.channel as TextChannel).bulkDelete(amount);
-      await slash.reply({
-        content: `Truncating ${amount} messages...`,
-        ephemeral: true,
-      });
+      if (amount < 1 || amount > 100) {
+        slash.reply({
+          content: "Amount must be between 1 and 100 !",
+          ephemeral: true,
+        });
+      } else {
+        try {
+          await (slash.channel as TextChannel).bulkDelete(amount);
+          await slash.reply({
+            content: `Truncating ${amount} messages...`,
+            ephemeral: true,
+          });
+        } catch (error) {
+          slash.reply({
+            content:
+              "Message truncation failed ! Messages must be younger than 14 days !",
+            ephemeral: true,
+          });
+        }
+      }
     } catch (error) {
       slash.reply({
         content:
-          "Message truncation failed ! Messages must be younger than 14 days !",
+          "Message truncation failed ! Something is wrong with your request... i hope you had rode the description before using the command !",
         ephemeral: true,
       });
     }
